requirements.txt
pillow
pandas
scikit-image
opencv-python
numpy
matplotlib